Adds support for partial data files
Adds support for data files for included partials by passing all data from non-used files into the context of the base file as supplemental data (using the files path as the object key). This allows the use of the data either by assigning the data context of the included partial with a `lookup` helper or by using the supplemental data in a `with` block.