Creates bash executables that can be used to configure and run `fswatch` on the server
This adds two bash scripts to help with configuring and running `fswatch` on the server:

- The `watch` script configures and starts a series of `fswatch` tasks in the background and subsequently calls the `index` script each time a file change is detected. Changes to the templating engine's `engine/` and/or `patterns/` folder will trigger a reindexing of all sites whereas changes specific to a site's domain root folder (e.g., `libraries.emory.edu/`) and/or `data/` folder with only trigger a reindexing for the given site.

- The `index` script neatly handles the reindexing of one or more sites based on the level at which a file change was detected. That is, if an environment-level file change was detected, it will compile a list of all available sites and force a reindexing for each one. Otherwise, if a site-level file change was detected, it will only force a reindexing of the given site.

One thing to note is that the `watch` script will need to be updated to register new domains as additional sites go live. In terms of setup on the server, though, this now allows us to simply execute our `watch` script one time via a shell at the server's root to properly configure a collection of `fswatch` tasks across all sites and environments.