Reconfigures the index to implement a true API and fixes index file corruption
This reconfigures the indexer to implement a true API and adds a `GET /status` endpoint (in addition to the known `POST /index` endpoint) to query the indexer for `READY` or `INDEXING` status. This also updates the Cache to detect file corruption, and reset the cache if/when detected. Lastly, this also updates the engine's API to skip caching if/when an index process is running, which should eliminate the issue of trying to write to the cache while it may be in the process of updating.